Avoid using the quote character (") in our chunk encoding.
This quote character requires escaping in JSON strings, and using
it for BLOCK_DIRT was causing the average chunk to be somewhere on
the order of ~50kb. This patch corrects that problem by increasing
the lower number to above that of a quotation mark, reducing the
average chunk size back to ~33kb. Ideally we will apply compression
and base64 in the future for even more space savings, but that's a
bit more involved (see https://github.com/crazy2be/compression-tests)