Catch pre-boot (i.e. probably requirejs) errors and display them to the user.
This lets the user know that things are broken, and that they should probably
not bother waiting any longer. I was originally planning to use fatalError for
this, but then we would need to have a require statement which loaded
fatalError (and *just* fatalError, lest there was a load error for one of the
other modules), assigns window.onError, and then loads the remaining modules
required for boot-up. And this still wouldn't catch, say, an error when the
fatalError module itself is loaded, or when requirejs is loaded. This solution
catches all such errors immediately.

Ideally such errors should never occur in production anyway, and if they *do*,
it means we've done something really wrong.

Fixes #142