Quick workaround for #147.
The old code used to totally destroy it's state when it got a
prediction it didn't recognize, so it had to throw or else things
would break in stranger ways. However, a while back the code was
rewritten to be a bit more robust, so that throwing was no longer
necessary, and we could recover more gracefully. This patch takes
advantage of that to avoid crashing the client.

I wouldn't really consider this a "fix" for #147 per se, because
in order to consider it as such, we shouldn't give a warning under
normal operation of the server, but should still be robust if the
server code gets a little bit broken. So it's a workaround.