Add some benchmarker for the meshers
We want to make sure the meshing code is as fast as possible, but previously
it was basically impossible to determine how fast it was actually running,
or see if a change made a difference. This commit adds some simple
benchmarking code so we can tell when the meshing becomes faster or slower
after a change.

Right now it takes ~50ms to mesh a simple flat chunk on my computer. We can
probably get this down a fair bit.