Remove center vertex, speeding up meshing by ~15%.
The node profiler tools didn't seem to think this function took very much
time, so I didn't even touch it before, but it seems to have much more
of an effect than any of the other functions the profiler thought were
a big deal. Makes the profiler kind of useless, but I digress.