Add deployment notes for nearlyfreespeech
The deployment process here involves a couple of steps and is
a bit tricky to figure out; it seems worthwhile to try and note
the manual steps required for future redeployments.

This document is intended to allow for other deployment
hosts / targets to be added as well, in case such proves useful.